Move/refactor cloneSelection into a mixin
